add check for srcbmi and cleanup
